test: make sure seed only runs on playwright init and not for individual tests

+++ src/seed/test-data/index.ts
@@ -14,6 +14,11 @@ export const seedTestData = async (payload: Payload): Promise<void> => {
     return;
   }
 
+  // see comment in playwright be config
+  if (isPlaywright && process.env.DOSEED !== 'true') {
+    return;
+  }
+
   try {
     // delete all collections
     await deleteData(payload);
